+++ compact_str/src/features/mod.rs
@@ -18,6 +18,8 @@ mod garde;
 mod markup;
 #[cfg(feature = "proptest")]
 mod proptest;
+#[cfg(feature = "pyo3")]
+mod pyo3;
 #[cfg(feature = "quickcheck")]
 mod quickcheck;
 #[cfg(feature = "rkyv")]